Update UploadTask.java
By allowing the fileKey to be null, we can choose whether we want to send the file by itself (e.g. to S3) or to use a standard form body which we are still using.
Since Android 11 and up use contentResolvers, the code has been updated to use a content resolver to load the file and also to resolve the mime type.
The Multipart body was refactored into a helper method to clean up the code a bit.
The default method was changed back to POST.